[refactor] Move GitHub integration into new subfolder
Note: Function renderPrDialog() in prosemirror-lwdita/command.ts is technically a function only used for the GitHub integration, but it will stay in the root folder.
Because otherwise we would need to create an import in the prosemirror-demo package in module demo-plugin.ts, which would result in a dependency from the demo to the Github Integration!
Therefore it can stay in the main package root folder.
