+++ packages/prosemirror-lwdita/src/index.ts
@@ -21,6 +21,6 @@ export * from './plugin';
 export * from './commands';
 export * from './untravel-document';
 export * from './attributes';
-export * from './request';
-export * from './toast';
-export * from './github.plugin';
+export * from './github-integration/request';
+export * from './github-integration/toast';
+export * from './github-integration/github.plugin';
