+++ packages/prosemirror-lwdita/tests/github.plugin.spec.ts
@@ -16,7 +16,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
 import { expect } from 'chai';
-import { fetchRawDocumentFromGitHub, transformGitHubDocumentToProsemirrorJson } from '../src/github.plugin';
+import { fetchRawDocumentFromGitHub, transformGitHubDocumentToProsemirrorJson } from '../src/github-integration/github.plugin';
 import fetchMock from 'fetch-mock';
 import { shortXdita, shortXditaProsemirroJson } from './test-utils';
